Fix plushangup not being recognized.
The iwar.conf file that comes with iwar sets two options called "plushangup" and "plushangupsleep". The old version of the code checked for "PlusHangup" and "PlusHangupsleep". This change brings the name of the options in line with the others. You could also modify the provided iwar.conf if you'd prefer.